requirements.txt
opencv-python
scikit-learn
keras
tensorflow
pillow
matplotlib
